enemy movement logic update (queen movement)
updated targeting array so that all entities can use it instead of just the player
changed movement controller to an autoload
changed enemy ai so that they spawn the movement targets for a queen card and chooses the closest available cell to the player. This all happens in one frame so the enemy targets aren't visible to the player.

Some of this code might be spaghetti.