gold update
The player's inventory component now has a gold variable, and each enemy has a new gold stack item that contains a certain amount of gold. Upon death, any entity (including the player) will drop whatever is in their inventory and that item will spawn on the map on top of the carcass.