options screen implemented
options screen with volume slider and mute button for both music and sfx.
to_do: apply a nice theme to it.
Currently thinking about other options to add to the screen. If no other options to add, may change the button to a speaker icon and just call it the audio option screen.