Display Update and Card Store Fix
Made a DisplayManager autoload that checks that your screen resolution isn't smaller than the default project resolution (mine was), and if it is, set's the project resolution to slightly smaller than your screen

This update broke the card store display, so I fixed that and added a vertical scroll to prevent it from expanding past the top of the screen. Clicking store cards is still broken, working on fixing that next.