New title screen art and movement fix
I made an ascii version of the art for the title screen
I fixed the movement and backtracing, next up, implement the choose landing area function specifically for knight backtracing.
Warning: I commented out the item pickup because I was getting an error plus we have no items to pickup. I'll fix it later if and when it's needed.